+++ jaziku/utils/settings_run.py
@@ -529,8 +529,9 @@ def check():
     else:
         if not os.path.isfile(globals_vars.config_run["path_to_file_var_I"]):
             console.msg_error_configuration('path_to_file_var_I',
-                _("Can't open file '{0}' for var I, \nif you want run var I with internals files of jaziku\n"
-                  "you need set 'PATH TO FILE VAR I' as 'internal'").format(globals_vars.config_run["path_to_file_var_I"]))
+                _("Can't open file '{0}' for var I, \nplease check filename and check path relative (to runfile) or\n"
+                  "absolute to var I. If you want run var I with internals files\n"
+                  "of jaziku you need set 'PATH TO FILE VAR I' as 'internal'").format(globals_vars.config_run["path_to_file_var_I"]))
 
     # -------------------------------------------------------------------------
     # thresholds var_I
@@ -658,7 +659,7 @@ def query_yes_no(question, default="yes"):
 
 def continue_run():
 
-    if not globals_vars.args.force:
+    if not globals_vars.ARGS.force:
 
         query = query_yes_no(_("\nPlease check the configuration to run, continue?"))